Payment To Farmers Added

+++ src/main/java/com/dairy/dto/paymentToFarmer/PaymentToFarmerRequestDto.java
@@ -12,21 +12,17 @@ import lombok.NoArgsConstructor;
 public class PaymentToFarmerRequestDto {
 	
 	
-
 	private Date invoice_date;
 	private Date from_date;
 	private Date to_date;
-	private Long amount;
-
+	private Float amount;
 	private Long farmer;
 
 	private int branch;
-	
 	private Long total_collected_milk;
-	private String milktype;
-	
-	private Long feed_deduction;
-	private Long advance_deduction;
+	private String milktype;	
+	private Float feed_deduction;
+	private Float advance_deduction;
 	private String payment_method;
 	private String payment_note;
